added initial implementation for shader reload, moved loading of shaders to loadShader function, fixed warnings, updated readme

+++ gg/src/graphics/Renderer.cpp
@@ -69,7 +69,7 @@ namespace gg
 			}
 			case Shader::SystemUniform::TIME:
 			{
-				_mat->setUniformf(Shader::systemUniformEnumToString(*it), Time::getCurrentTime());
+				_mat->setUniformf(Shader::systemUniformEnumToString(*it), static_cast<float>(Time::getCurrentTime()));
 				break;
 			}
 			}
